Some more work on the collision broadphase

+++ Core/Source/Physics/GPUCollisionGrid.cpp
@@ -1,7 +1,7 @@
 #include "GPUCollisionGrid.h"
 #include <cassert>
 
-Mupfel::GPUCollisionGrid::GPUCollisionGrid(size_t in_num_cells_x, size_t in_num_cells_y, size_t in_entities_per_cell, size_t in_cell_size_pow) :
+Mupfel::GPUCollisionGrid::GPUCollisionGrid(uint32_t in_num_cells_x, uint32_t in_num_cells_y, uint32_t in_entities_per_cell, uint32_t in_cell_size_pow) :
 	num_cells_x(in_num_cells_x), num_cells_y(in_num_cells_y), EntitiesPerCell(in_entities_per_cell), cell_size_pow(in_cell_size_pow)
 {
 	assert(is_powerof2(num_cells_x));
